requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv

-e .
